Event Page and cart functionality

+++ Backend/ViewModels/AddProductModel.cs
@@ -11,7 +11,7 @@ public class AddProductModel
     public IFormFile? Image3 { get; set; }
     public IFormFile? Image4 { get; set; }
     public IFormFile? Image5 { get; set; }
-    public required decimal Price { get; set; }
+    public required float Price { get; set; }
     public string? CategoryId { get; set; }
     public List<string>? Tags { get; set; }
 }
